Fix user_data to install git before enabling WinRM
The user_data and remote_exec currently get kicked off at the
same time. The remote exec retries until WinRM is enabled.

Previously user_data was enabling WinRM before installing git and
creating the profile to add it to the path.

In this PR I've also changed how the path is set to use a
powershell profile.